Add a health endpoint
Implement a HTTP api at '/' and '/health' which returns success.

This is a Passive health check which means it does not check if
the other parts are running as expected. It simply returns with
HTTP 200 OK status everytime the endpoint is called.

Also runs kube_monkey() function in background with go routines.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>